Refactorizado el código de la interfaz para que la parte del juego no esté en el login
